requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.18
anthropic>=0.42.0
google-genai>=1.0.0
openai>=1.60.0
gspread>=6.1.0
google-auth>=2.37.0
python-dotenv>=1.0.1
